Move stringify.js to a more common location for reuse.
