+++ lib/drivers/abbott/cli/fslibre.js
@@ -13,7 +13,7 @@ import pkg from '../../../../package.json';
 import builder from '../../../objectBuilder';
 import abbottFreeStyleLibre from '../abbottFreeStyleLibre';
 
-import stringify from './stringify';
+import stringify from '../../common/stringify';
 
 // eslint-disable-next-line no-underscore-dangle
 global.__DEBUG__ = true;
