+++ lib/drivers/common/stringify.js
